in progress check-in of data service work

+++ src/main/java/io/reign/data/DataValue.java
@@ -33,6 +33,6 @@ public interface DataValue<T> {
      * 
      * @return time to live in millis
      */
-    public long getTimeToLive();
+    public int getTimeToLive();
 
 }
